Add EnrollStudentsData class
