+++ src/main/java/teammates/ui/output/StudentsData.java
@@ -12,7 +12,6 @@ import teammates.common.datatransfer.attributes.StudentAttributes;
 public class StudentsData extends ApiOutput {
 
     private List<StudentData> students;
-    private List<EnrollError> failToEnrollStudents;
 
     public StudentsData() {
         this.students = new ArrayList<>();
@@ -20,11 +19,6 @@ public class StudentsData extends ApiOutput {
 
     public StudentsData(List<StudentAttributes> students) {
         this.students = students.stream().map(StudentData::new).collect(Collectors.toList());
-    }
-
-    public StudentsData(List<StudentAttributes> students, List<EnrollError> failToEnrollStudents) {
-        this.students = students.stream().map(StudentData::new).collect(Collectors.toList());
-        this.failToEnrollStudents = failToEnrollStudents;
     }
 
     public List<StudentData> getStudents() {
@@ -35,11 +29,4 @@ public class StudentsData extends ApiOutput {
         this.students = students;
     }
 
-    public List<EnrollError> getFailToEnrollStudents() {
-        return failToEnrollStudents;
-    }
-
-    public void setFailToEnrollStudents(List<EnrollError> failToEnrollStudents) {
-        this.failToEnrollStudents = failToEnrollStudents;
-    }
 }
