+++ src/web/app/pages-instructor/instructor-search-page/instructor-search-page.component.spec.ts
@@ -52,7 +52,6 @@ describe('InstructorSearchPageComponent', () => {
         sectionName: 'Section 2',
       },
     ],
-    failToEnrollStudents: [],
   };
 
   beforeEach(() => {
